Catch delete of history table so a crashed process doesn't leave a mess
In the case that the process that create a history metric crashes
it is unable to delete the metric from folsom and re-create it.
This addresses that. Ideally a process should be started when a
history is created, it should trap exists and recreate the
history table in the face of a crash. I'll do that later, honest.