Use anonymous table for history (like histogram reservoir)
In multi-app environments (like riak) naming stats with a 2 tuple
is useful. It makes querying stats by app e.g. {riak_kv, _} much simpler.
History metrics use the name of the metric as the name of the data table,
this commit uses an anonympus table and stores a record in the
?HISTORY_TABLE. Since that table is already defined and created
maybe this was the original intent?