Fix handling of c_compiler_used when only a major version is provided
This mostly occurs when erlang is compiled under MSC since it only
provides a single number as the version.  GCC can trigger this issue if it
does not have a minor version, but is probably rare in the wild.

Additionally, this allows for the case where erlang wasn't able to
determine which compiler was used (i.e. {undefined, undefined}).

Add test for the handling of c_compiler_used by convert_system_info