Make spiral and slide_uniform histogram metrics faster
There are 2 places in folsom where an ets:insert_new is done immediately
followed by an ets:update_counter on the same key. Since, in the normal
case, the key is likely to already exist, this can be optimized by
trying the update_counter first in a try/catch and only do the
insert_new if needed. This improves performance of spiral and
slide_uniform by about 13% in simple benchmarks.