Let the slide server crash and stay crashed after multiple errors
The restart strategy on the slide server was such that it would
_never_ stop in the case that the ets table it was trimming went away.
Instead it would metronimically add a crash message to your log
every 30 seconds. This commit sets the restart strategy to 3
failures in 3 minutes. Since the only seen live examples of
a slide server crash have been when the ets table has gone
away, it is best crash and stay dead than be restarted after
3 failed trims.