Fix a race against querying the ETS/DETS tables.
There is a window from the query of, say, `ets:all()` to the point where we fetch out the table data. The table may disappear in between. This patch fixes this problem by ignoring such tables and returning the empty list for those where there is no data available anymore.