Add new_spiral/2
This adds new_spiral(Name, Update) where update is one of `fast` or
`no_exceptions`.  `fast` works by catching the `badarg` on update of a
non-existant key.  `no_exceptions` instead reads the ETS table first
to avoid catching an exception.  The reason for this is that if one
uses a spiral in a cowboy onresponse function, if there's an exception
thrown by the handler it'll be overwritten by folsom.  Maybe there's
another way around this?

TODO: tests